Refactor: libraries folder name
