+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handlers } from '@/libraries/nextAuth';
+import { handlers } from '@/libraries/next-auth';
 
 export const { GET, POST } = handlers;
