+++ src/libraries/oracledb/whitelist/service.ts
