+++ src/libraries/oracledb/whitelist/sql.ts
